+++ tests/unit/test_cache_keys.py
@@ -62,7 +62,9 @@ def test_create_key__normalize_duplicate_params():
         method='GET', url='https://img.site.com/base/img.jpg?param_1=a&param_1=b&k=v'
     )
     request_2 = Request(
-        method='GET', url='https://img.site.com/base/img.jpg?param_1=b&param_1=a', params={'k': 'v'}
+        method='GET',
+        url='https://img.site.com/base/img.jpg?param_1=b&param_1=a',
+        params={'k': 'v'},
     )
     assert create_key(request_1) == create_key(request_2)
 
@@ -86,27 +88,26 @@ def test_redact_response__escaped_params():
         assert redacted_response.raw.url == 'https://img.site.com/base/img.jpg?where=code%3D123'
 
 
-def test_normalize_request__json_body():
+@pytest.mark.parametrize(
+    'content_type',
+    ['application/json', 'application/json; charset=utf-8'],
+)
+@pytest.mark.parametrize(
+    'data',
+    [
+        b'{"param_1": "value_1", "param_2": "value_2"}',
+        b'["param_3", "param_2", "param_1"',
+    ],
+)
+def test_normalize_request__json_body(data, content_type):
     request = Request(
         method='GET',
         url='https://img.site.com/base/img.jpg',
         data=b'{"param_1": "value_1", "param_2": "value_2"}',
-        headers={'Content-Type': 'application/json'},
+        headers={'Content-Type': content_type},
     )
     norm_request = normalize_request(request, ignored_parameters=['param_2'])
     assert norm_request.body == b'{"param_1": "value_1", "param_2": "REDACTED"}'
-
-
-def test_normalize_request__json_body_list():
-    """Support request body with a list as a JSON root"""
-    request = Request(
-        method='GET',
-        url='https://img.site.com/base/img.jpg',
-        data=b'["param_3", "param_2", "param_1"]',
-        headers={'Content-Type': 'application/json'},
-    )
-    norm_request = normalize_request(request)
-    assert norm_request.body == b'["param_1", "param_2", "param_3"]'
 
 
 def test_normalize_request__json_body_list_filtered():
@@ -140,17 +141,21 @@ def test_normalize_request__json_body_empty():
     assert normalize_request(request, ignored_parameters=['param_2']).body == b'{}'
 
 
-def test_normalize_request__binary_body():
+@pytest.mark.parametrize(
+    'content_type',
+    ['application/octet-stream', None],
+)
+def test_normalize_request__binary_body(content_type):
     request = Request(
         method='GET',
         url='https://img.site.com/base/img.jpg',
         data=b'some bytes',
-        headers={'Content-Type': 'application/octet-stream'},
+        headers={'Content-Type': content_type},
     )
     assert normalize_request(request, ignored_parameters=['param']).body == request.data
 
 
-def test_normalize_request__ovsersized_body():
+def test_normalize_request__oversized_body():
     body = {'param': '1', 'content': '0' * MAX_NORM_BODY_SIZE}
     encoded_body = json.dumps(body).encode('utf-8')